+++ Sources/game/screen_items/TitleScreen.swift
@@ -8,8 +8,6 @@ struct TitleScreen {
 	private nonisolated(unsafe) var endY: Int { Screen.rows }
 
 	private nonisolated(unsafe) var selectedOptionIndex = 0
-	private nonisolated(unsafe) var selectedSettingOptionIndex = 0
-	private nonisolated(unsafe) var config = Config()
 
 	mutating func show() async -> TitleScreenOptions {
 		// Quickly start a new game in debug mode
@@ -73,7 +71,7 @@ struct TitleScreen {
 			}
 		}
 
-		func action(screen: inout TitleScreen) async {
+		func action() async {
 			switch self {
 				case .newGameOption:
 					await newGame()
@@ -89,110 +87,10 @@ struct TitleScreen {
 				case .helpOption:
 					return
 				case .settingsOption:
-					await screen.settings()
+					return
 				case .quitOption:
 					endProgram()
 			}
 		}
-	}
-
-	enum SettingsScreenOptions: CaseIterable {
-		case useNerdFontOption
-		case wasdKeysOption
-		case arrowKeysOption
-		case vimKeysOption
-
-		var label: String {
-			switch self {
-				case .useNerdFontOption:
-					"Use Nerd Font"
-				case .vimKeysOption:
-					"Use Vim Keys for moving"
-				case .arrowKeysOption:
-					"Use Arrow Keys for moving"
-				case .wasdKeysOption:
-					"Use WASD Keys for moving"
-			}
-		}
-
-		func action(config: inout Config) {
-			switch self {
-				case .useNerdFontOption:
-					config.useNerdFont.toggle()
-				case .vimKeysOption:
-					config.vimKeys.toggle()
-				case .arrowKeysOption:
-					config.arrowKeys.toggle()
-				case .wasdKeysOption:
-					config.wasdKeys.toggle()
-			}
-		}
-
-		func configOption(_ config: Config) -> String {
-			switch self {
-				case .useNerdFontOption:
-					config.useNerdFont ? "On" : "Off"
-				case .vimKeysOption:
-					config.vimKeys ? "On" : "Off"
-				case .arrowKeysOption:
-					config.arrowKeys ? "On" : "Off"
-				case .wasdKeysOption:
-					config.wasdKeys ? "On" : "Off"
-			}
-		}
-	}
-
-	mutating func settings() async {
-		// TODO: reload config after saving so this isn't async
-		config = await Config.load()
-		let text = "Settings"
-		let x = middleX
-		let y = middleY - (Screen.rows / 2)
-		while true {
-			Screen.clear()
-			Screen.print(x: x - (text.count / 2), y: y, text.styled(with: .bold))
-			var lastIndex = SettingsScreenOptions.allCases.count - 1
-			var yStart = 3
-			for (index, option) in SettingsScreenOptions.allCases.enumerated() {
-				yStart = await printSettingsOption(x: x, y: y + yStart, index: index, text: option.label, configOption: option.configOption(config))
-				lastIndex = index
-			}
-
-			let skip = lastIndex + 1
-			yStart = await printSettingsOption(x: x, y: yStart + 1, index: lastIndex + 2, text: "Save and Quit", configOption: "")
-			yStart = await printSettingsOption(x: x, y: yStart, index: lastIndex + 3, text: "Quit", configOption: "")
-			let key = TerminalInput.readKey()
-			switch key {
-				case .up, .w, .k, .back_tab:
-					selectedSettingOptionIndex = max(0, selectedSettingOptionIndex - 1)
-					if selectedSettingOptionIndex == skip {
-						selectedSettingOptionIndex = selectedSettingOptionIndex - 1
-					}
-				case .down, .s, .j, .tab:
-					selectedSettingOptionIndex = min(SettingsScreenOptions.allCases.count - 1 + 3, selectedSettingOptionIndex + 1)
-					if selectedSettingOptionIndex == skip {
-						selectedSettingOptionIndex = selectedSettingOptionIndex + 1
-					}
-				case .enter:
-					if selectedSettingOptionIndex == lastIndex + 2 {
-						await config.write()
-						await Game.shared.loadConfig()
-						return
-					} else if selectedSettingOptionIndex == lastIndex + 3 {
-						return
-					} else {
-						SettingsScreenOptions.allCases[selectedSettingOptionIndex].action(config: &config)
-					}
-				default:
-					break
-			}
-		}
-	}
-
-	private func printSettingsOption(x: Int, y: Int, index: Int, text: String, configOption: String) async -> Int {
-		let isSelected = selectedSettingOptionIndex == index
-		let configOptionToPrint = configOption == "" ? "" : ": \(configOption)"
-		await Screen.print(x: x - (text.count / 2), y: y, "\(isSelected ? "\(Game.shared.config.icons.selectedIcon) " : " ")\(text)\(configOptionToPrint)".styled(with: .bold, styledIf: isSelected))
-		return y + 1
 	}
 }
